리뷰 좌표 타입 수정
- bugfix: 좌표 입력 시 타입 long -> double 변경

+++ src/main/java/com/bbdn/server/domain/interfaces/request/PlaceReviewUploadRequest.java
@@ -16,8 +16,8 @@ public class PlaceReviewUploadRequest {
     private Long placeId; // 장소ID
     private String addressName; // 전체도로명주소
     private String roadAddress; // 도로명주소
-    private Long x; // x좌표
-    private Long y; // y좌표
+    private double x; // x좌표
+    private double y; // y좌표
 
     private String reviewMainContent; // 리뷰 개요
     private String reviewGoodContent; // 장점 리뷰 내용
